src/test/java/org/asciicerebrum/neocortexengine/mechanics/conditionevaluators/impl/NotEvaluatorTest.java
package org.asciicerebrum.neocortexengine.mechanics.conditionevaluators.impl;

import org.asciicerebrum.neocortexengine.domain.core.ICharacter;
import org.asciicerebrum.neocortexengine.domain.core.UniqueEntity;
import org.asciicerebrum.neocortexengine.domain.game.DndCharacter;
import org.asciicerebrum.neocortexengine.domain.game.Weapon;
import org.asciicerebrum.neocortexengine.mechanics.conditionevaluators.ConditionEvaluator;
import org.junit.After;
import org.junit.AfterClass;
import static org.junit.Assert.assertFalse;
import static org.junit.Assert.assertTrue;
import org.junit.Before;
import org.junit.BeforeClass;
import org.junit.Test;
import static org.mockito.Mockito.mock;
import static org.mockito.Mockito.when;

/**
 *
 * @author species8472
 */
public class NotEvaluatorTest {

    private NotEvaluator evaluator;

    private ConditionEvaluator conditionEvaluator;

    public NotEvaluatorTest() {
    }

    @BeforeClass
    public static void setUpClass() {
    }

    @AfterClass
    public static void tearDownClass() {
    }

    @Before
    public void setUp() {
        this.evaluator = new NotEvaluator();
        this.conditionEvaluator = mock(ConditionEvaluator.class);

        this.evaluator.setConditionEvaluator(this.conditionEvaluator);
    }

    @After
    public void tearDown() {
    }

    @Test
    public void evaluateFalseTest() {
        final ICharacter dndCharacter = new DndCharacter();
        final UniqueEntity contextItem = new Weapon();
        when(this.conditionEvaluator.evaluate(dndCharacter, contextItem))
                .thenReturn(Boolean.TRUE);
        final boolean result = this.evaluator.evaluate(
                dndCharacter, contextItem);
        assertFalse(result);
    }

    @Test
    public void evaluateTrueTest() {
        final ICharacter dndCharacter = new DndCharacter();
        final UniqueEntity contextItem = new Weapon();
        when(this.conditionEvaluator.evaluate(dndCharacter, contextItem))
                .thenReturn(Boolean.FALSE);
        final boolean result = this.evaluator.evaluate(
                dndCharacter, contextItem);
        assertTrue(result);
    }
}
